Quality of life fixes; also if the map was smaller than the screen it used to make the scale 0 due to integer division, I changed it by explicitly casting one of the values to float.

+++ app/src/main/java/ca/uoguelph/pspenler/maptracker/MapImageView.java
@@ -268,7 +268,7 @@ public class MapImageView extends android.support.v7.widget.AppCompatImageView {
 
         mImageWidth = displayMetrics.widthPixels;
         mImageHeight = Math.round(mImageWidth * aspectRatio);
-        imageScale = mBitmap.getWidth() / mImageWidth;
+        imageScale =  ((float) mBitmap.getWidth()) / mImageWidth;
         translateY = prevTranslateY = (displayMetrics.heightPixels / 2) - (mImageHeight / 2) + 60;
 
         invalidate();
